+++ CHANGES.rst
@@ -88,6 +88,7 @@ OpenREM version history
 * `#449`_  Exports: Fixed a bug that prevented fluoro exports if protocol names had non-ASCII characters
 * `#448`_  Documentation: Added a diagram showing the relationship between the OpenREM system components
 * `#447`_  Imports: Modified rdsr and ctdetail template to import and display data from Pixelmed generated Toshiba RDSR
+* `#446`_  Import: Extract additional Philips private information for Allura Xper systems, create workaround for missing end angles for rotational acquisitions
 * `#445`_  Interface: Added function for user to determine between DX and fluoro for ambiguous modalities
 * `#444`_  Imports: DX systems that submit RDSRs that look like fluoro can now be reclassified using `#445`_
 * `#443`_  Exports: Accession number and ID are now exported to XLSX as text. Thanks to `@LuukO`_